Re-word sections regarding puisne committee members
I am not necessarily attached to the term "General Member of the Management Committee", and am open to other suggestions for a suitable wording change.

The term "Puisne" and other removed/updated section on this topic are a potential source of confusion for members reading the constitution. For example, (former) section 7.4 is redundant considering there are multiple Puisne Members, and only adds confusion as to their duties and role within the management committee. This change aims to clarify the general purpose of the Puisne Members of the Management Committee, such that a layman reading the constitution or attending a General Meeting will better understand the purpose of the Puisne Members.

Finally, I will note that the term "Puisne" is already considered obsolete in many jurisdictions.