Rework temporary appointments to Management Committee
Section 10.2 currently allows the Management Committee to make arbitrary long-term appointments of Officers of the Management Committee. Consider the case where a T3 member falls ill early in their term and resigns; their replacement can be appointed by the Management Committee without a General Meeting, and their appointment can extend the full year to the next Annual General Meeting.

As Officers are those with significant privilege and responsibility, these appointments should be limited as practical to ensure the Members of the Society have a chance to vote on their appointment. Additionally, we extend the ability for Management Committee to appoint Puisne Members to fill vacancies, under the same restrictions as appointing Officers. Note that by only extending this to positions vacated by resignation, arbitrary removals and appointments to "pack" the Management Committee should not be a concern.