Additional qualifiers for the dissolution clause
In the event of dissolution, the entire remaining wealth of the Society is transferred to a single association with similar purposes. This should be allowed to be split between multiple societies, in the event that no single association exactly fulfills the same purpose, and should be ideally targeted towards associations aimed at students of the University of Queensland - but not exclusively targeted, in case no suitable association exists for UQ students.